[FIX] controllers and models updated a lot

+++ pong/app/controllers/api/guild_members_controller.rb
@@ -4,21 +4,26 @@ module Api
 
     def index
       guild = Guild.find(params[:guild_id])
-      render json: guild, status: :ok
+      render json: guild.members, status: :ok
     end
 
     def update
-      guild_member = GuildMember.find(params[:id])
-      if guild_member.role == 'master' || params[:role] == 'master' || params[:role] == 2
-        render json: { error: 'master is immutable role.' }, status: :bad_request
+      guild_member = GuildMember.find_by(user_id: params[:user_id])
+      problem = GuildMember.update_check(current_user, guild_member, params[:role])
+      if problem['status'].nil?
+        guild_member.update!(role: params[:role])
+        render json: guild_member, status: :ok
+      else
+        render json: { message: problem['message'] }, status: problem['status']
       end
-      guild_member.role = params[:role]
-      guild_member.save!
-      render json: guild_member, status: :no_content
     end
 
     def destroy
-      guild_member = GuildMember.find(params[:id])
+      guild_member = GuildMember.find_by(user_id: params[:user_id])
+      unless GuildMember.can_destroy?(current_user, guild_member)
+        render json: { message: '' }, status: :forbidden and return
+      end
+
       if guild_member.role == 'master'
         guild_member.guild.destroy!
       else
